Vulkan Wrapper!
This code is designed to wrap up lots of the vulkan functionality I will be using in the future.  The refactor is not done yet. The only thing left right now is support for textures on models.   There is also lots of redundant code that will be removed in later submissions.