Setting up deferred renderer
This changelist includes code that will be the basis for the deferred renderer.  I've also added the render texture class, whose responsibility will be to set up a texture which will be used to be rendered to.  Aside from this, I've change some code to follow my chosen convention, mostly this involved renaming variables and functions to conform to the same style.

There are bugs in this code, but I've decided this is good stopping point because the changelist was getting too large.