Compute Shaders and Pipelines
This changelist will introduce architectural changes to accommodate for the need of compute shaders in the renderer.  Materials have now being split into visual materials and compute materials, along with a material_base class which has common functions to both.   Also introduced is the compute pipeline and graphics pipeline; their requirements and complexities are different so they need to be separated.  Tying them together is the pipeline base class.

Changes to the material store have been implemented to support the two types of materials now present.

If also added a compute family queue for the purposes of running compute shaders.

Lastly, I've added a compute shader for testing to see if it compiles and loads into the material store c++ code.