Compute shader in place
I am now able to run compute shaders by using the compute  pipeline.  I've also added a compute shader file which at the moment does nothing, is just there as a placeholder, but eventually it will house the code that generates 3D textures.

The material_base class now only has code that is common for both visual materials and compute materials.

The rest of the changes have to do with making the code follow the conventions I have adopted with naming variables.