Investigating voxel texture not completely rendering
This submission will make sure that the voxel texture renders completely, however, it uses _device->wait_for_all_operations_to_finish function.  At this point, I know there is a synchronization issue happening in the voxelization set of commands, but it is not clear to me exactly how to fix.